Add files via upload
This program nreceives an input and a matrix
�
x
�
nxnMade in order
 Takes a number to assign a value to each of the array's houses. A snail from home(0,0)
( 0 ,0 )It starts and goes around the matrix in a spiral and clockwise manner until it reaches the innermost point of the matrix. In this way, the snail collects numbers every house it moves forward. If this sum was a perfect square, it is worth one point for him, and we want the sum of these points in the output of the program.